refactor(effect): split slotScreensChanged, and document PR #905
The two remaining items from the PR #906 audit.

slotScreensChanged was one 775-line body carrying five phases, and it was the
direct cause of two comment desyncs the audit had to fix. The removed-screens
pass is the bulk of it, and its two arms reach opposite conclusions about the
same event: a screen leaves the managed set either because the DESKTOP changed,
where the windows are still owned by that desktop's live autotile session and
must only be demoted, or because autotile was genuinely turned off for it, where
nothing owns them any more and they are untracked and restored. Those are now
demoteWindowsForDesktopSwitch and untrackWindowsForDisabledScreens, and the slot
keeps the epoch bookkeeping and the managed-set write around them.

Both bodies moved verbatim: stripped of whitespace they compare byte-identical
to the originals, so this is behaviour-preserving by construction rather than by
inspection. Both take the caller's window snapshot and append to its deferred
windowed-fullscreen release list rather than releasing inline, because that
release has to land after the managed-set write.

The extraction pushed signals.cpp to 1253 lines, past the ceiling, so the whole
concern moved to tilinghandler/screenschanged.cpp. signals.cpp is 420 lines and
the new file is 862.

CHANGELOG: PR #905 merged into this cycle with no entry. Written from its four
commits rather than its summary, covering what a user sees, which is scrolling
mode no longer holding GPU time and video memory for columns off the viewport,
smaller backdrop captures for the blur family, and a daemon idle reclaim that
now actually reclaims.

Build green, 358/358 tests pass, no warnings.
