+++ CHANGELOG.md
@@ -57,6 +57,7 @@ Versioning follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 - **Fixed-size windows no longer jump to the top of the screen while scrolling off**: a scrolling column window that cannot fill its column, like a game launcher with a fixed size, is committed centered in the column, whether its size was constrained as it was placed or PlasmaZones centered it once the application settled on its size. The paint that keeps a column visible while it scrolls out of view was given the column's position instead of the distance to move it, so it drew the window at the column's top left corner and the window leapt to the top of the screen the moment it left the viewport, riding there until the scroll settled. It now leaves and arrives at the centered position it actually occupies, and a size-constrained window no longer gets a spurious slide out of that corner every time the strip moves ([#906](https://github.com/fuddlesworth/PlasmaZones/pull/906)).
+- **Scrolling mode no longer spends GPU time and video memory on columns you cannot see**: a column scrolled out of the viewport stayed fully alive in the render pipeline. Its decoration fold kept running at up to sixty times a second, and it held roughly 136 MB of full-canvas render targets per 4K window, so a couple of windows in scrolling mode could sit at high GPU load and hundreds of megabytes of video memory with nothing on screen to account for it. Off-viewport columns are now skipped by the compositor, and a column parked for more than ten seconds releases its render targets outright, while scrolling back to one that was parked briefly is still a cheap redraw. Backdrop captures for the blur family are taken at the resolution their samplers actually read, which cuts them from about 33 MB to about 2 MB per 4K window. The daemon's five second idle reclaim also works now, where before it queued the release behind a render pass that never came ([#905](https://github.com/fuddlesworth/PlasmaZones/pull/905)).
 - **Snap Assist keeps showing window thumbnails**: after a snap or two the window picker fell back to program icons and stayed on them until the daemon was restarted. The daemon clears its thumbnail cache half a minute after the picker closes, and the KWin effect went on skipping the windows it believed were still cached, so nothing ever re-captured them. The two sides now agree and the picker shows thumbnails every time. Several capture faults are fixed alongside it. A window with no drawable frame yet no longer posts an invisible thumbnail, a small window is no longer enlarged to thumbnail size and blurred, an offscreen render no longer leaks a texture per capture, and a thumbnail that fails to load now falls back to the program icon instead of leaving an empty square. Clicking and hovering a card also stays reliable while thumbnails are still arriving, because an arriving thumbnail no longer rebuilds the whole picker. The experimental zero-copy path (PLASMAZONES_DMABUF_THUMBNAILS) had its own form of the same fault from the second picker onward, and its thumbnails arrived upside down. Both are fixed ([#902](https://github.com/fuddlesworth/PlasmaZones/pull/902)).
 - **A window that declares a too-large minimum size after opening now floats**: a scrolling column window was floated at its native size when it opened too large for the screen, but an application that raised its minimum size only after opening stayed tiled at a size the strip could never satisfy. Some games running under Wine do exactly that, opening small and then pinning their minimum to the configured game resolution. Such a window now leaves the strip and floats the moment its minimum outgrows the screen, the same answer it would have gotten at open. Unfloating it later returns it to the column it left ([#900](https://github.com/fuddlesworth/PlasmaZones/pull/900)).
 - **Tab indicators stay behind the windows in front of them**: a tabbed scrolling column's indicator was drawn above every window on the screen, so a floating window raised over the strip wore a stripe of the column behind it. The indicator now sits above the columns it belongs to and behind anything stacked over the strip, the way it would if it were an ordinary window ([#896](https://github.com/fuddlesworth/PlasmaZones/pull/896)).
