+++ kwin-effect/CMakeLists.txt
@@ -111,6 +111,7 @@ set(plasmazones_effect_SRCS
     tilinghandler/minimizefloat.cpp
     tilinghandler/outputchange.cpp
     tilinghandler/signals.cpp
+    tilinghandler/screenschanged.cpp
     tilinghandler/state.cpp
     tilinghandler/tiling.cpp
     tilinghandler/wiring.cpp
